+++ packages/benchmark-js/src/index.ts
@@ -1,3 +1,4 @@
 export * from './Benchmark';
 export * from './BenchmarkGroup';
 export { BenchmarkLoggerLevel } from './Logger';
+export * from './types';