+++ packages/benchmark-js/src/types.internal.ts
@@ -50,3 +50,5 @@ export interface BenchmarkStats {
      */
     variance: number;
 }
+
+export type Hrtime = [number, number];